Support ArmMetadata with API version 2022-09-01
The new format of ArmMetadata found here
https://management.azure.com/metadata/endpoints?api-version=2022-09-01
