+++ src/Authentication.Abstractions.Test/TestData/ArmResponse2022-09-01.json
@@ -1,0 +1,41 @@
+{
+  "portal": "https://portal.azure.com",
+  "authentication": {
+    "loginEndpoint": "https://login.microsoftonline.com",
+    "audiences": [ "https://management.core.windows.net/", "https://management.azure.com/" ],
+    "tenant": "common",
+    "identityProvider": "AAD"
+  },
+  "media": "https://rest.media.azure.net",
+  "graphAudience": "https://graph.windows.net/",
+  "graph": "https://graph.windows.net/",
+  "name": "AzureCloud",
+  "suffixes": {
+    "azureDataLakeStoreFileSystem": "azuredatalakestore.net",
+    "acrLoginServer": "azurecr.io",
+    "sqlServerHostname": "database.windows.net",
+    "azureDataLakeAnalyticsCatalogAndJob": "azuredatalakeanalytics.net",
+    "keyVaultDns": "vault.azure.net",
+    "storage": "core.windows.net",
+    "azureFrontDoorEndpointSuffix": "azurefd.net",
+    "storageSyncEndpointSuffix": "afs.azure.net",
+    "mhsmDns": "managedhsm.azure.net",
+    "mysqlServerEndpoint": "mysql.database.azure.com",
+    "postgresqlServerEndpoint": "postgres.database.azure.com",
+    "mariadbServerEndpoint": "mariadb.database.azure.com",
+    "synapseAnalytics": "dev.azuresynapse.net",
+    "attestationEndpoint": "attest.azure.net"
+  },
+  "batch": "https://batch.core.windows.net/",
+  "resourceManager": "https://management.azure.com/",
+  "vmImageAliasDoc": "https://raw.githubusercontent.com/Azure/azure-rest-api-specs/master/arm-compute/quickstart-templates/aliases.json",
+  "activeDirectoryDataLake": "https://datalake.azure.net/",
+  "sqlManagement": "https://management.core.windows.net:8443/",
+  "microsoftGraphResourceId": "https://graph.microsoft.com/",
+  "appInsightsResourceId": "https://api.applicationinsights.io",
+  "appInsightsTelemetryChannelResourceId": "https://dc.applicationinsights.azure.com/v2/track",
+  "attestationResourceId": "https://attest.azure.net",
+  "synapseAnalyticsResourceId": "https://dev.azuresynapse.net",
+  "logAnalyticsResourceId": "https://api.loganalytics.io",
+  "ossrDbmsResourceId": "https://ossrdbms-aad.database.windows.net"
+}
